// frontend/src/app/model/bus.model.ts
export interface Bus {
    _id?: string; 
    operatorName: string;
    busType: string;
    departureTime: string;
    rating: number[];
    totalSeats?: number;
    routes: string; 
    images: string;
    liveTracking: number;
    reschedulable: number;
    driverContact?: string; 
    bimage: string;
    fare: number; 
  }